% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkG.R
\name{checkG}
\alias{checkG}
\title{Check for consistency in genetic effects}
\usage{
checkG(pcgen.output, suffStat, alpha = 0.01, covariates = NULL)
}
\arguments{
\item{pcgen.output}{a graph with nodes G (genotype) and a number of traits.
Typically output from \code{pcgen} or \code{pcgenFast}.}

\item{suffStat}{A data.frame with \eqn{(p+1)} columns, the first of which is
the factor G (genotype, i.e samples' IDs) and subsequent columns contain
the traits, and optionally some QTLs. The name of the first column should
be G. Should not contain covariates.}

\item{alpha}{The significance level used in each conditional independence
test. Default is 0.01. The test itself of course does not depend on this,
but it is used in the EM-algorithm to speed up calculations. More
precisely, the EM-algorithm is stopped once the P value is below the
significance level. This can be done because the PC algorithm only needs an
accept/ reject decision.}

\item{covariates}{A data.frame containing covariates that should always be
used in each conditional independence test. Should be either \code{NULL}
(default) or a data.frame with the same number of rows as \code{suffStat}.
An intercept is already included for each trait in suffStat; covariates
should not contain a column of ones.}
}
\value{
A logical matrix of dimension \eqn{(p+1) \times (p+1)}, \eqn{p} being
the number of traits. Most entries are \code{FALSE}, except those in the
first row and column for which there are conflicts.  Entries \eqn{[1, j]}
and \eqn{[j, 1]} are \code{TRUE} if the \eqn{j}th trait has significant
genetic variance, but there is no partially directed path from G towards
that trait. The matrix can then be used in a subsequent run of \code{pcgen}
or \code{pcgenFast}, in the \code{fixedEdges} argument. The arguments
\code{suffStat}, \code{alpha} and \code{covariates} should stay the same
throughout (first run of \code{pcgen}, \code{checkG}, second run of
\code{pcgen}).
}
\description{
Given output from \code{\link{pcgen}} or \code{\link{pcgenFast}}, this
function checks whether the estimated graph is consistent with the set of
traits having significant genetic variance. The function detects traits that
have significant genetic variance but for which there is no partially
directed path from G.
}
\references{
Kruijer, W., Behrouzi, P., Bustos-Korts, D., Rodríguez-Álvarez,
M. X., Mahmoudi, S. M., Yandell, B., ... & van Eeuwijk, F. A. (2020).
Reconstruction of networks with direct and indirect genetic effects.
\emph{Genetics}, 214(4), 781-807.
}
