% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTriple2.R
\name{checkTriple2}
\alias{checkTriple2}
\title{param a, b, c, nbrsA, nbrsC, sepsetA, sepsetC, version.unf, maj.rule, verbose:
param suffStat data.frame, of which the first column is the factor genotype,
param alpha (Default 0.01) The significance level used in the test. The test
param QTLs column numbers in suffStat that correspond to QTLs
param K (Default NULL) The kinship (i.e Genetic Relationship Matrix)
param covariates data.frame containing covariates, that should always be used
param Vg, Ve (Default NULL) Genetic and residual covariance matrices of
param dec (Default NULL) Contains a spectral decomposition of K = Z Z^t.}
\usage{
checkTriple2(
  a,
  b,
  c,
  nbrsA,
  nbrsC,
  sepsetA,
  sepsetC,
  suffStat,
  alpha,
  version.unf = c(NA, NA),
  maj.rule = FALSE,
  verbose = FALSE,
  covariates = NULL,
  QTLs = integer(),
  K = NULL,
  max.iter = 50,
  stop.if.significant = TRUE,
  use.res = FALSE,
  res.cor = NULL
)
}
\description{
param a, b, c, nbrsA, nbrsC, sepsetA, sepsetC, version.unf, maj.rule, verbose:
param suffStat data.frame, of which the first column is the factor genotype,
param alpha (Default 0.01) The significance level used in the test. The test
param QTLs column numbers in suffStat that correspond to QTLs
param K (Default NULL) The kinship (i.e Genetic Relationship Matrix)
param covariates data.frame containing covariates, that should always be used
param Vg, Ve (Default NULL) Genetic and residual covariance matrices of
param dec (Default NULL) Contains a spectral decomposition of K = Z Z^t.
}
