% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2traits.R
\docType{data}
\name{d2treps}
\alias{d2treps}
\title{Simulated dataset of a 2 traits network with replicates}
\format{
A data frame of dimension \eqn{ 1000 \times 3 }. The first column is
the factor G (genotype); the subsequent columns contain \eqn{Y_1} and
\eqn{Y_2}.
}
\usage{
data(d2treps)
}
\description{
Simulated data, for two replicates of 500 genotypes g1,...,g500. Two traits
were simulated (Y1 and Y2), such that G -> Y1 -> Y2
}
\seealso{
Other pcgen datasets: 
\code{\link{d2tsnpsmeans}},
\code{\link{simdata}}
}
\concept{pcgen datasets}
\keyword{datasets}
