% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2traits.R
\docType{data}
\name{d2tsnpsmeans}
\alias{d2tsnpsmeans}
\title{Simulated dataset of a 2 traits network with replicates means}
\format{
A data frame of dimension \eqn{ 500 \times 103}. The first column is
the factor G (genotype); then \eqn{Y_1} and \eqn{Y_2}, followed by the 100
markers \code{mxxxx} where \code{xxx} is the marker number
}
\usage{
data(d2treps)
}
\description{
The means of 100 markers and two traits corresponding to the dataset
\code{d2treps} of 500 genotypes g1,...,g500, where G -> Y1 -> Y2
}
\seealso{
Other pcgen datasets: 
\code{\link{d2treps}},
\code{\link{simdata}}
}
\concept{pcgen datasets}
\keyword{datasets}
