% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gemmaGK}
\alias{gemmaGK}
\title{R interface to 'Genome-wide Efficient Mixed Model Association' (GEMMA)}
\usage{
gemmaGK(
  genoinputs,
  gk,
  miss = 0.05,
  maf = 0.01,
  r2 = 0.9999,
  hwe = 0,
  notsnp = FALSE,
  gxe = "",
  outprefix = "out",
  outdir = "output",
  license = FALSE
)
}
\arguments{
\item{gk}{Integer variable for calculating the relatedness matrix. A value
of 1 (the default) generates a centered matrix. A value of 2 generates a
standardized matrix}
}
\description{
GEMMA is an efficient association testing software of Linear Mixed Models
(LMM)s and related GWAS models. This interface is for the Relatedness matrix
calculation functionality in centered or standardized matrix form
}
