% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencovTest.R
\name{gencovTest}
\alias{gencovTest}
\title{Estimate genetic covariances between all pairs of traits, and test their
significance}
\usage{
gencovTest(suffStat, max.iter = 200, out.cor = TRUE)
}
\arguments{
\item{suffStat}{A data.frame with \eqn{(p+1)} columns, the first of which is the factor G
(genotype, i.e samples' IDs) and subsequent columns contain the traits, and
optionally some QTLs. The name of the first column should be G. Should not
contain covariates.}

\item{max.iter}{Maximum number of iterations in the EM-algorithm, used to fit
the bivariate mixed model (when \code{use.res = FALSE)}.}

\item{out.cor}{If \code{TRUE}, the output will contain estimates of genetic
correlations; otherwise covariances. The pvalues are always for genetic
covariance.}
}
\value{
A list with elements pvalues and out.cor, which are both p x p
matrices
}
\description{
This will perform the conditional independence test used in pcgen algorithm,
assuming there are replicates, and K = Z Z^t
}
\examples{
\dontrun{
data(simdata)
test <- gencovTest(suffStat= simdata, max.iter = 200, out.cor= TRUE )
}

}
\references{
Kruijer, W., Behrouzi, P., Bustos-Korts, D., Rodríguez-Álvarez,
M. X., Mahmoudi, S. M., Yandell, B., ... & van Eeuwijk, F. A. (2020).
Reconstruction of networks with direct and indirect genetic effects.
\emph{Genetics}, 214(4), 781-807.
}
\author{
Willem Kruijer and Pariya Behrouzi. Maintainers: Willem Kruijer
\email{willem.kruijer@wur.nl} and Pariya Behrouzi
\email{pariya.behrouzi@gmail.com}
}
