% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.cons.intern2.R
\name{pc.cons.intern2}
\alias{pc.cons.intern2}
\title{param sk, version.unf, maj.rule, verbose: as in the original pc.cons.intern function
param suffStat data.frame, of which the first column is the factor genotype,
param alpha (Default 0.01) The significance level used in the test. The test
param QTLs column numbers in suffStat that correspond to QTLs
param K (Default NULL) The kinship (i.e Genetic Relationship Matrix)
param covariates data.frame containing covariates, that should always be used}
\usage{
pc.cons.intern2(
  sk,
  suffStat,
  alpha,
  version.unf = c(NA, NA),
  maj.rule = FALSE,
  verbose = FALSE,
  covariates = NULL,
  QTLs = QTLs,
  K = NULL,
  max.iter = 50,
  stop.if.significant = TRUE,
  use.res = FALSE,
  res.cor = NULL
)
}
\description{
param sk, version.unf, maj.rule, verbose: as in the original pc.cons.intern function
param suffStat data.frame, of which the first column is the factor genotype,
param alpha (Default 0.01) The significance level used in the test. The test
param QTLs column numbers in suffStat that correspond to QTLs
param K (Default NULL) The kinship (i.e Genetic Relationship Matrix)
param covariates data.frame containing covariates, that should always be used
}
