% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgen.R
\name{pcgen}
\alias{pcgen}
\title{The pcgen algorithm}
\usage{
pcgen(
  suffStat,
  covariates = NULL,
  QTLs = integer(),
  K = NULL,
  alpha = 0.01,
  m.max = Inf,
  fixedEdges = NULL,
  fixedGaps = NULL,
  verbose = FALSE,
  use.res = FALSE,
  res.cor = NULL,
  max.iter = 50,
  stop.if.significant = TRUE,
  NAdelete = FALSE,
  return.pvalues = FALSE
)
}
\arguments{
\item{suffStat}{data.frame, of which the first column is the factor genotype
(i.e samples' IDs) and subsequent columns contain the traits.
The name of the first column should be G
Should not contain covariates.}

\item{covariates}{data.frame containing covariates that should always be used
in each conditional independence test. Should be either NULL
(default) or a data.frame with the same number of rows as
suffStat}

\item{QTLs}{column numbers in suffStat data.frame that correspond to QTLs.
These may be partly in S and x and y, but x and y cannot be both QTLs.}

\item{K}{The genetic relatedness matrix. If NULL (the default), independent
genetic effects are assumed.}

\item{alpha}{(Default 0.01) The significance level used in each conditional
independence test.}

\item{m.max}{Maximum size of the conditioning sets.}

\item{fixedEdges}{A logical matrix of dimension (p+1) * (p+1), where
p is the number of traits. The first row and column
refer to the node genotype, and subsequent rows and
columns to the traits. As in the pcalg package,
the edge i-j is never considered for removal
if the entry \link{i,j} or \link{j,i} (or both) are TRUE.
In that case, the edge is guaranteed to be present
in the resulting graph.}

\item{fixedGaps}{A logical matrix of dimension (p+1) * (p+1), where
p is the number of traits. The first row and column
refer to the node genotype, and subsequent rows and
columns to the traits. As in the pcalg package,
the edge i-j is removed before starting the algorithm
if the entry \link{i,j} or \link{j,i} (or both) are TRUE.
In that case, the edge is guaranteed to be absent
in the resulting graph.}

\item{verbose}{If TRUE, p-values for the conditional independence
tests are printed}

\item{use.res}{If TRUE, the test for conditional independence of 2 traits given
a set of other traits and G is based on residuals from GBLUP.
If FALSE (the default), it is based on bivariate mixed models.}

\item{res.cor}{Correlation matrix of the residuals of the GBLUP. In case
it is NULL (default), the test for conditional independence
between Yj and Yk given a set of traits YS is based on a
bivariate mixed model. If res.cor is provided, it is based
on partial correlations among the residuals, tested using
the gaussCItest function from pcalg. The residuals can be
obtained using the function getResiduals}

\item{max.iter}{Maximum number of iterations in the EM-algorithm.}

\item{stop.if.significant}{If TRUE, the EM-algorithm used in some
of the conditional independence tests will be stopped
whenever the p-value becomes significant, i.e. below
alpha. This will speed up calculations, and can be done
because (1) the PC algorithm only needs an accept/reject
decision (2) In EM the likelihood is nondecreasing. Should
be put to FALSE if the precise p-values are of interest.}

\item{NAdelete}{Logical, and is the NAdelete option in pcalg::pc function: TRUE
deletes the edge corresponding to the conditional independence
test when it returns NA. Default is FALSE}

\item{return.pvalues}{If FALSE, the output is a graph (an object with S3 class
\code{"pcgen"}). If TRUE, the output is a list with
elements \code{gr} (the graph) and \code{pMax} (a matrix
with the the maximal p-value for each edge).}
}
\value{
a graph (an object with S3 class \code{"pcgen"})
}
\description{
The pcgen algorithm, assuming independent genetic effects.
}
\details{
(1) This is an adaptation of the pc function from pcalg, with a
different conditional independence test (see pcgenTest), as well as
modified orientation rules  (2) Some parameters from the original
pc function are fixed here: skel.method = "stable", u2pd = "relaxed",
conservative = FALSE, maj.rule = TRUE and solve.confl = TRUE, while the user
can set the "NAdelete" parameter (defaults to FALSE).
Labels (defining the names of the nodes of the
graph) is derived from the data-frame suffStat, containing the data
(3) pcgen requires phenotypic observations on multiple traits,
measured on the same samples. The current implementation extends the genetic
relatedness matrix definition from K = Z Z^t, Z being the incidence matrix
assigning plants to genotypes (as is common in plants data) to a
a marker-based relatedness matrix definition (appropriate for human data).
(4) the test for conditional independence between Yj and Yk given
a set of traits YS can be based either on a bivariate mixed model
(put res.cor = NULL) or on partial correlations among the residuals.
In the latter case, res.cor should be the correlation matrix of the
residuals, obtained with getResiduals.
}
\references{
Kruijer et al. (2020) Reconstruction of networks with direct and
indirect genetic effects
}
\author{
Willem Kruijer and Pariya Behrouzi.
Maintainers: Willem Kruijer \email{willem.kruijer@wur.nl} and
Pariya Behrouzi \email{pariya.behrouzi@gmail.com}
}
