% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgen2.R
\docType{package}
\name{pcgen2}
\alias{pcgen2}
\title{pcgen2: A package for reconstructing genetic traits networks}
\description{
The pcgen2 package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

