% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcRes.R
\name{pcRes}
\alias{pcRes}
\title{The pc algorithm applied to residuals}
\usage{
pcRes(
  suffStat,
  alpha = 0.01,
  K = NULL,
  m.max = Inf,
  verbose = FALSE,
  covariates = NULL,
  QTLs = integer(),
  cov.method = "uni",
  use.GBLUP = FALSE,
  return.pvalues = FALSE
)
}
\arguments{
\item{suffStat}{A data.frame, of which the first column is the factor G
(genotype, i.e samples' IDs) and subsequent columns contain the traits, and
optionally some QTLs. The name of the first column should be G. Should not
contain covariates.}

\item{alpha}{The significance level used in each conditional independence
test. Default is 0.01}

\item{K}{kinship}

\item{m.max}{maximum size of the conditioning set, in the pc-algorithm on the residuals}

\item{verbose}{If \code{TRUE}, p-values for the conditional independence
tests are printed}

\item{covariates}{A data.frame containing covariates that should always be
used in each conditional independence test. Should be either \code{NULL}
(default) or a data.frame with the same number of rows as \code{suffStat}.
An intercept is already included for each trait in suffStat; covariates
should not contain a column of ones.}

\item{QTLs}{Column numbers in \code{suffStat} that correspond to QTLs. These
may be partly in S and x and y, but x and y cannot be both QTLs.}

\item{cov.method}{(Default 'fe') A string, specifying which method should be
used to compute the G-BLUP. Options are 'us'
(unstructured multi-trait model fitted
using sommer) and uni' (based on univariate GBLUPs).}

\item{use.GBLUP}{Use the GBLUP itself, instead of the residuals (as in Topner et al)}

\item{return.pvalues}{}
}
\value{
A graph (an object with S3 class \code{"pcgen"})
}
\description{
The pc algorithm applied to residuals, assuming there are replicates,
and K = Z Z^t.
}
\details{
Here we
put a lot
of details
COMMENT
}
\references{
A paper on arxiv and Topner et al (2017)
}
\author{
Willem Kruijer and Pariya Behrouzi.
Maintainers: Willem Kruijer \email{willem.kruijer@wur.nl} and
Pariya Behrouzi \email{pariya.behrouzi@gmail.com}
}
