% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton2.R
\name{skeleton2}
\alias{skeleton2}
\title{param suffStat data.frame, of which the first column is the factor genotype,
param alpha (Default 0.01) The significance level used in the test. The test
param labels, p, method, m.max, fixedGaps, fixedEdges, NAdelete, verbose: as
in the original skeleton function
param covariates data.frame containing covariates, that should always be used
param QTLs column numbers in suffStat that correspond to QTLs
param K (Default NULL) The kinship (i.e Genetic Relationship Matrix)}
\usage{
skeleton2(
  suffStat,
  alpha,
  labels,
  p,
  method = c("stable", "original"),
  m.max = Inf,
  fixedGaps = NULL,
  fixedEdges = NULL,
  NAdelete = TRUE,
  verbose = FALSE,
  covariates = NULL,
  QTLs = integer(),
  K = NULL,
  dec = NULL,
  max.iter = 50,
  stop.if.significant = TRUE,
  use.res = FALSE,
  res.cor = NULL
)
}
\description{
param suffStat data.frame, of which the first column is the factor genotype,
param alpha (Default 0.01) The significance level used in the test. The test
param labels, p, method, m.max, fixedGaps, fixedEdges, NAdelete, verbose: as
in the original skeleton function
param covariates data.frame containing covariates, that should always be used
param QTLs column numbers in suffStat that correspond to QTLs
param K (Default NULL) The kinship (i.e Genetic Relationship Matrix)
}
