% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udag2pdagRelaxed2.R
\name{udag2pdagRelaxed2}
\alias{udag2pdagRelaxed2}
\title{param non.collider.nodes (Default 1) A vector of integers}
\usage{
udag2pdagRelaxed2(
  gInput,
  verbose = FALSE,
  unfVect = NULL,
  solve.confl = FALSE,
  orientCollider = TRUE,
  rules = rep(TRUE, 3),
  non.collider.nodes = 1
)
}
\description{
param non.collider.nodes (Default 1) A vector of integers
}
