% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkG.R
\name{checkG}
\alias{checkG}
\title{Check for consistency in genetic effects}
\usage{
checkG(pcgen.output, suffStat, alpha = 0.01, covariates = NULL)
}
\arguments{
\item{pcgen.output}{a graph with nodes G (genotype) and a number of
traits. Typically output from pcgen or pcgenFast}

\item{suffStat}{data.frame, of which the first column is the factor genotype
(i.e samples' IDs) and subsequent columns contain the traits.
The name of the first column should be G
Should not contain covariates.}

\item{alpha}{(Default 0.01) The significance level used in each conditional
independence test.}

\item{covariates}{data.frame containing covariates that should always be used
in each conditional independence test. Should be either NULL
(default) or a data.frame with the same number of rows as
suffStat}
}
\value{
A logical matrix of dimension (p+1) x (p+1), p being the number of traits.
Most entries are FALSE, except those in the first row and column for which there
are conflicts.
Entries \link{1,j} and \link{j,1} are TRUE if the jth trait has
significant genetic variance, but there is no
partially directed path from G towards that trait. The matrix
can then be used in a subsequent run of pcgen or pcgenFast, in the
fixedEdges argument. The arguments suffStat, alpha and covariates
should stay the same throughout
(first run of pcgen, checkG, second run of pcgen).
}
\description{
Given output from pcgen or pcgenFast, this function checks whether the
estimated graph is consistent with the set of traits having significant genetic variance.
The function detects traits that have significant genetic variance but for which there is no
partially directed path from G.
}
\details{
Details
}
\references{
A paper on arxiv
}
\author{
Willem Kruijer and Pariya Behrouzi.
Maintainers: Willem Kruijer \email{willem.kruijer@wur.nl} and
Pariya Behrouzi \email{pariya.behrouzi@gmail.com}
}
