% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{Residuals from GBLUP}
\usage{
getResiduals(suffStat, covariates = NULL, cov.method = "uni", K = NULL)
}
\arguments{
\item{suffStat}{data.frame, of which the first column is the factor genotype
(i.e samples' IDs) and subsequent columns contain the traits.
The name of the first column should be G
Should not contain covariates.}

\item{covariates}{data.frame containing covariates that should always be used
in each conditional independence test. Should be either NULL
(default) or a data.frame with the same number of rows as
suffStat}

\item{cov.method}{(Default 'uni') A string, specifying whether the BLUP of the genetic effects
a multi-trait mixed model should be fitted
using sommer ('us'), or univariate mixed models ('uni')}

\item{K}{An optional marker-based relatedness matrix of dimension n x n, n being the
number of unique genotypes in the first column in suffStat. In case suffStat contains
replicates, the resulting relatedness of the observations will be Z K Z^t, where Z
is the incidence matrix assigning plants to genotypes}
}
\value{
A data-frame with the residuals
}
\description{
Computes residuals from the BLUP (best linear unbiased predictor) of the genetic effects
}
\details{
The residuals are used in pcgenFast and pcRes
}
\references{
A paper on arxiv
}
\author{
Willem Kruijer and Pariya Behrouzi.
Maintainers: Willem Kruijer \email{willem.kruijer@wur.nl} and
Pariya Behrouzi \email{pariya.behrouzi@gmail.com}
}
