% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgenFast.R
\name{pcgenFast}
\alias{pcgenFast}
\title{The pcgenFast algorithm}
\usage{
pcgenFast(
  suffStat,
  alpha = 0.01,
  m.max = Inf,
  res.m.max = Inf,
  verbose = FALSE,
  covariates = NULL,
  fixedEdges = NULL,
  QTLs = integer(),
  max.iter = 50,
  stop.if.significant = TRUE,
  cov.method = "uni",
  use.res = FALSE,
  return.pvalues = FALSE
)
}
\arguments{
\item{suffStat}{data.frame, of which the first column is the factor genotype
(i.e samples' IDs) and subsequent columns contain the traits.
The name of the first column should be G
Should not contain covariates.}

\item{alpha}{(Default 0.01) The significance level used in each conditional
independence test.}

\item{m.max}{maximum size of the conditioning set, in pcgen}

\item{res.m.max}{maximum size of the conditioning set, in the pc-algorithm on the residuals}

\item{verbose}{If TRUE, p-values for the conditional independence
tests are printed}

\item{covariates}{data.frame containing covariates that should always be used
in each conditional independence test. Should be either NULL
(default) or a data.frame with the same number of rows as
suffStat}

\item{fixedEdges}{A logical matrix of dimension (p+1) * (p+1), where
p is the number of traits. The first row and column
refer to the node genotype, and subsequent rows and
columns to the traits. As in the pcalg package,
the edge i-j is never considered for removal
if the entry \link{i,j} or \link{j,i} (or both) are TRUE.
In that case, the edge is guaranteed to be present
in the resulting graph.}

\item{QTLs}{column numbers in suffStat data.frame that correspond to QTLs.
These may be partly in S and x and y, but x and y cannot be both QTLs.}

\item{max.iter}{Maximum number of iterations in the EM-algorithm.}

\item{stop.if.significant}{If TRUE, the EM-algorithm used in some
of the conditional independence tests will be stopped
whenever the p-value becomes significant, i.e. below
alpha. This will speed up calculations, and can be done
because (1) the PC algorithm only needs an accept/reject
decision (2) In EM the likelihood is nondecreasing. Should
be put to FALSE if the precise p-values are of interest.}

\item{cov.method}{(Default 'uni') A string, specifying which method should be
used to estimate the covariance structure required for the
computation of conditional means, as well as for the
computation of residuals in the first step. Options are 'us' (unstructured multi-trait model fitted
using sommer), and uni' (based on univariate GBLUPs).}

\item{use.res}{If FALSE, residuals are only used for screening. If TRUE,
residuals are also used the
test for conditional independence of 2 traits given a set of
other traits and G}

\item{return.pvalues}{}
}
\value{
A graph (an object with S3 class \code{"pcgen"})
}
\description{
This will run the pcgen algorithm, assuming there are replicates,
and K = Z Z^t. Fast version, with residuals-based pre-selection.
}
\details{
Here we
put a lot
of details
COMMENT on cov.method (the same option is
used in the computation of the residuals and the conditional means)
fixedGaps and fixedEdges are NULL
m.max, alpha,.... are also used in the first step
}
\references{
A paper on arxiv
}
\author{
Willem Kruijer and Pariya Behrouzi.
Maintainers: Willem Kruijer \email{willem.kruijer@wur.nl} and
Pariya Behrouzi \email{pariya.behrouzi@gmail.com}
}
