% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgenTest.R
\name{pcgenTest}
\alias{pcgenTest}
\title{The conditional independence test in pcgen}
\usage{
pcgenTest(
  x,
  y,
  S,
  suffStat,
  covariates = NULL,
  QTLs = integer(),
  K = NULL,
  alpha = 0.01,
  max.iter = 50,
  stop.if.significant = TRUE,
  use.res = FALSE,
  res.cor = NULL
)
}
\arguments{
\item{x, y}{Column numbers in \code{suffStat} that should be tested for
conditional independence given the variables in \code{S}.}

\item{S}{vector of integers defining the conditioning set, where the integers
refer to column numbers in \code{suffStat}. May be numeric(), i.e. the
empty set.}

\item{suffStat}{A data.frame with \eqn{(p+1)} columns, the first of which is the
factor G (genotype, i.e samples' IDs) and subsequent columns contain the traits,
and optionally some QTLs. The name of the first column should be G. Should not
contain covariates.}

\item{covariates}{A data.frame containing covariates that should always be
used in each conditional independence test. Should be either \code{NULL}
(default) or a data.frame with the same number of rows as \code{suffStat}.
An intercept is already included for each trait in suffStat; covariates
should not contain a column of ones.}

\item{QTLs}{Column numbers in \code{suffStat} that correspond to QTLs. These may be
partly in \eqn{S} and \eqn{x} and \eqn{y}, but \eqn{x} and \eqn{y} cannot be both QTLs.
Note: the factor genotype (column number 1) may occur in S, as well as x and y}

\item{K}{The kinship (i.e genetic relatedness matrix . If NULL (the default), independent
genetic effects are assumed. Azza: conside the def below originally in getResiduals.R:}

\item{alpha}{The significance level used in each conditional independence
test. Default is 0.01. The test itself of course does not depend on this,
but it is used in the EM-algorithm to speed up calculations. More
precisely, the EM-algorithm is stopped once the P value is below the
significance level. This can be done because the PC algorithm only needs an
accept/ reject decision.}

\item{max.iter}{Maximum number of iterations in the EM-algorithm, used to fit
the bivariate mixed model (when \code{use.res = FALSE)}.}

\item{stop.if.significant}{If TRUE, the EM-algorithm used in some of the
conditional independence tests will be stopped whenever the p-value becomes
significant, i.e. below alpha. This will speed up calculations, and can be
done because (1) the PC algorithm only needs an accept/reject decision (2)
In EM the likelihood is nondecreasing. Should be put to FALSE if the
precise p-values are of interest.}

\item{use.res}{If \code{TRUE}, the test for conditional independence of 2
traits given a set of other traits and \eqn{G} is based on residuals from
GBLUP. If \code{FALSE} (the default), it is based on bivariate mixed
models.}

\item{res.cor}{If \code{use.res = TRUE}, \code{res.cor} should be the
correlation matrix of the residuals of the GBLUP- the residuals can be
obtained using the \code{getResiduals} function. If \code{res.cor} is
provided, it is based on partial correlations among the residuals, tested
using the gaussCItest function from pcalg. In case \code{res.cor} is
\code{NULL} (the default), the test for conditional independence between Yj
and Yk given a set of traits YS is based on a bivariate mixed model.}
}
\value{
a p-value
}
\description{
This performs the conditional independence test used in the pcgen algorithm,
assuming there are replicates, and independent genetic effects.
}
\details{
This will perform the conditional independence test used in pcgen algorithm,
assuming (replicates and K = Z Z^t; or no replicates with a generic kinship
matrix K = A) and (independent genetic effects)

\code{pcgenTest} tests for conditional independence between \eqn{x} and
\eqn{y} given \eqn{S}. It distinguishes 2 situations: (i) if one of \eqn{x}
and \eqn{y} (say \eqn{x}) is the factor G, \code{pcgenTest} will test if the
genetic variance in \eqn{y} is zero, given the traits in S. (ii) if \eqn{x}
and \eqn{y} are both traits, \code{pcgenTest} tests if the residual
covariance between them is zero, given the traits in \eqn{S} and the factor
G. The factor G is automatically included in the conditioning set \eqn{S}
(\eqn{S} does not need to contain the integer 1). This test is either based
on a bivariate mixed model (when \code{use.res=FALSE}), or on residuals from
GBLUP (\code{use.res=T}), obtained with the getResiduals function. In the
latter case, \code{res.cor} must be provided.
}
\references{
\itemize{
\item Kruijer, W., Behrouzi, P., Bustos-Korts, D., Rodríguez-Álvarez,
M. X., Mahmoudi, S. M., Yandell, B., ... & van Eeuwijk, F. A. (2020).
Reconstruction of networks with direct and indirect genetic effects.
\emph{Genetics}, 214(4), 781-807.
}
}
\seealso{
\code{\link{getResiduals}}, \code{\link[pcalg]{gaussCItest}},
\code{\link[pcalg]{disCItest}}, \code{\link[pcalg]{binCItest}}

@examples
\dontrun{
 data(simdata)
 rs <- getResiduals(suffStat= simdata)
 pcgenTest(suffStat= simdata, x= 2, y= 3, S= 4)
 pcgenTest(suffStat= simdata, x= 2, y= 3, S= c(1,4))
 pcgenTest(suffStat= simdata, x= 2, y= 3, S= 4, use.res= TRUE, res.cor= cor(rs))
 pcgenTest(suffStat= simdata, x= 2, y= 1, S= 4)
 }
}
\author{
Willem Kruijer and Pariya Behrouzi. Maintainers: Willem Kruijer
\email{willem.kruijer@wur.nl} and Pariya Behrouzi
\email{pariya.behrouzi@gmail.com}
}
