% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.covar.test.R
\name{res.covar.test}
\alias{res.covar.test}
\title{EXACT test for residual correlation, based on a bivariate mixed model}
\usage{
res.covar.test(
  x,
  y,
  G,
  Z.t = NULL,
  K = NULL,
  X = as.data.frame(matrix(0, length(x), 0)),
  alpha = 0.01,
  use.manova = TRUE,
  max.iter = 50,
  stop.if.significant = TRUE
)
}
\arguments{
\item{x, y}{trait vectors whose independence is to be tested, conditional on
G (genotype) and possible covariates in X. Manadatory}

\item{G}{Factor of genotypes. Manadatotry}

\item{Z.t}{Defaults to \code{NULL}}

\item{K}{The kinship (i.e genetic relatedness matrix) . If \code{NULL} (the default),
independent genetic effects are assumed.}

\item{X}{A dataframe of possibly the covariates, traits and QTLs used as a
conditioning set. If not specified, this becomes a dependence test}

\item{alpha}{The significance level used in each conditional independence
test. Default is 0.01. The test itself of course does not depend on this,
but it is used in the EM-algorithm to speed up calculations. More
precisely, the EM-algorithm is stopped once the P value is below the
significance level. This can be done because the PC algorithm only needs an
accept/ reject decision.}

\item{use.manova}{}

\item{max.iter}{Maximum number of iterations in the EM-algorithm, used to fit
the bivariate mixed model (when \code{use.res = FALSE)}.}

\item{stop.if.significant}{If TRUE, the EM-algorithm used in some of the
conditional independence tests will be stopped whenever the p-value becomes
significant, i.e. below alpha. This will speed up calculations, and can be
done because (1) the PC algorithm only needs an accept/reject decision (2)
In EM the likelihood is nondecreasing. Should be put to FALSE if the
precise p-values are of interest.}
}
\description{
EXACT test for residual correlation, based on a bivariate mixed model
}
