% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton2.R
\name{skeleton2}
\alias{skeleton2}
\title{Infer the DAG Of a Phenotypic Traits' network with Genetic Effects}
\usage{
skeleton2(
  suffStat,
  alpha,
  labels,
  p,
  method = c("stable", "original"),
  m.max = Inf,
  fixedGaps = NULL,
  fixedEdges = NULL,
  NAdelete = TRUE,
  verbose = FALSE,
  covariates = NULL,
  QTLs = integer(),
  K = NULL,
  max.iter = 50,
  stop.if.significant = TRUE,
  use.res = FALSE,
  res.cor = NULL
)
}
\arguments{
\item{suffStat}{A data.frame with \eqn{(p+1)} columns, the first of which is the factor G
(genotype, i.e samples' IDs) and subsequent columns contain the traits, and
optionally some QTLs. The name of the first column should be G. Should not
contain covariates.}

\item{alpha}{The significance level used in each conditional independence
test. Default is 0.01. The test itself of course does not depend on this,
but it is used in the EM-algorithm to speed up calculations. More
precisely, the EM-algorithm is stopped once the P value is below the
significance level. This can be done because the PC algorithm only needs an
accept/ reject decision.}

\item{labels}{(optional) character vector of variable (or
    \dQuote{node}) names.  Typically preferred to specifying \code{p}.}

\item{p}{(optional) number of variables (or nodes).  May be specified
    if \code{labels} are not, in which case \code{labels} is set to
    \code{1:p}.}

\item{method}{Character string specifying method; the default,
    \code{"stable"} provides an \emph{order-independent} skeleton, see
    \sQuote{Details} below.}

\item{m.max}{Maximum size of the conditioning sets.}

\item{fixedGaps}{A logical matrix of dimension \eqn{(p+1) \times (p+1)},
where \eqn{p} is the number of traits. The first row and column refer to
the genotype node G, and subsequent rows and columns to the traits. As in
the pcalg package, the edge \eqn{i - j} is removed before starting the
algorithm if the entry \eqn{[i, j]} or \eqn{[j, i]} (or both) are
\code{TRUE}. In that case, the edge is guaranteed to be absent in the
resulting graph.}

\item{fixedEdges}{A logical matrix of dimension \eqn{(p+1) \times (p+1)},
where \eqn{p} is the number of traits. The first row and column refer to
the genotype node G, and subsequent rows and columns to the traits. As in
the pcalg package, the edge \eqn{i - j} is never considered for removal if
the entry \eqn{[i, j]} or \eqn{[j, i]} (or both) are \code{TRUE}. In that
case, the edge is guaranteed to be present in the resulting graph.}

\item{NAdelete}{Logical, and is the NAdelete option in pcalg::pc function:
TRUE deletes the edge corresponding to the conditional independence test
when it returns NA. Default is FALSE}

\item{verbose}{If \code{TRUE}, p-values for the conditional independence
tests are printed}

\item{covariates}{A data.frame containing covariates that should always be
used in each conditional independence test. Should be either \code{NULL}
(default) or a data.frame with the same number of rows as \code{suffStat}.
An intercept is already included for each trait in suffStat; covariates
should not contain a column of ones.}

\item{QTLs}{Column numbers in \code{suffStat} that correspond to QTLs. These
may be partly in S and x and y, but x and y cannot be both QTLs.
Note: the factor genotype (column number 1) may occur in S, as well as x and y}

\item{K}{The kinship (i.e genetic relatedness matrix . If NULL (the default), independent
genetic effects are assumed. Azza: conside the def below:#'}

\item{max.iter}{Maximum number of iterations in the EM-algorithm, used to fit
the bivariate mixed model (when \code{use.res = FALSE)}.}

\item{stop.if.significant}{If \code{TRUE}, the EM-algorithm used in some of
the conditional independence tests (when \code{use.res = FALSE}) will be
stopped whenever the p-value becomes significant, i.e. below \code{alpha}.
This will speed up calculations, and can be done because (1) the PC
algorithm only needs an accept/reject decision (2) In EM the likelihood is
nondecreasing. Should be put to \code{FALSE} if the precise p-values are of
interest.}

\item{use.res}{If \code{TRUE}, the test for conditional independence of 2
traits given a set of other traits and \eqn{G} is based on residuals from
GBLUP. If \code{FALSE} (the default), it is based on bivariate mixed
models.}

\item{res.cor}{If \code{use.res = TRUE}, \code{res.cor} should be the
correlation matrix of the residuals of the GBLUP- the residuals can be
obtained using the \code{getResiduals} function. If \code{res.cor} is
provided, it is based on partial correlations among the residuals, tested
using the gaussCItest function from pcalg. In case \code{res.cor} is
\code{NULL} (the default), the test for conditional independence between Yj
and Yk given a set of traits YS is based on a bivariate mixed model.}
}
\description{
Infer the DAG Of a Phenotypic Traits' network with Genetic Effects
}
\keyword{internal}
